chore: add tunnel unit test
